Move RFC3339 formatter into util.py
